+++ tournament-system/tournament-system-web/src/main/java/org/toursys/web/TournamentOverviewPage.java
@@ -330,12 +330,12 @@ public class TournamentOverviewPage extends BasePage {
                 private static final long serialVersionUID = 1L;
 
                 @Override
-                public Iterator<Groups> iterator(int first, int count) {
-                    return groups.subList(first, first + count).iterator();
+                public Iterator<Groups> iterator(long first, long count) {
+                    return groups.subList((int) first, (int) (first + count)).iterator();
                 }
 
                 @Override
-                public int size() {
+                public long size() {
                     return groups.size();
                 }
 
@@ -389,7 +389,7 @@ public class TournamentOverviewPage extends BasePage {
                 }
             };
             gridView.setRows(1);
-            gridView.setColumns(Math.max(1, dataProvider.size()));
+            gridView.setColumns((int) Math.max(1, dataProvider.size()));
             add(gridView);
         }
